refactored reference counting into common base class for both leveldb and eleveldb
